application/flask_transcriber/requirements.txt
Flask==3.0.3
supabase==2.5.1
python-dotenv==1.0.1
google-cloud-speech==2.26.0
google-generativeai==0.7.2
